Minor fixes
 - Fix compiler warnings

 - Improve camera panning on zoom

 - Fix bottom-right corner of map border missing

+++ Open-Rival/src/RenderUtils.cpp
@@ -26,16 +26,16 @@ namespace RenderUtils {
         return renderPos;
     }
 
-    int tileToScaledPx_X(int x, float zoomLevel) {
+    float tileToScaledPx_X(int x, float zoomLevel) {
         // Same as tileToPx_X, above, but with a zoom level applied
         float scaledTileWidth = tileSpriteWidthPx * zoomLevel;
         return x * (scaledTileWidth / 2);
     }
 
-    int tileToScaledPx_Y(int x, int y, float zoomLevel) {
+    float tileToScaledPx_Y(int x, int y, float zoomLevel) {
         // Same as tileToPx_Y, above, but with a zoom level applied
         float scaledTileHeight = tileSpriteHeightPx * zoomLevel;
-        int renderPos = y * (scaledTileHeight / 2);
+        float renderPos = y * (scaledTileHeight / 2);
         if (x % 2 == 1) {
             renderPos += (scaledTileHeight / 4);
         }
